Move things around to get ready for sharing
