+++ shared/settings/nav/index.native.js
@@ -1,0 +1,8 @@
+// @flow
+import type {Props} from './index'
+
+function SettingsNav (props: Props) {
+  return null
+}
+
+export default SettingsNav
