+++ shared/settings/payment/index.native.js
@@ -2,6 +2,7 @@
 import type {Props} from './index'
 
 function PaymentForm (props: Props) {
+  return null
 }
 
 export default PaymentForm
